ml/requirements.txt
Flask
tensorflow
numpy
Pillow
matplotlib
